Code Update V2
A major update to the code means that all the buttons are now 100% functional, with the addition of buttons that use PMOD for an external joystick.

In order to optimise the number of LUTs and provide a game that's fun to play but still challenging, we've made it so that the cars follow the frog on :
- The line above the car
- The line below the car
- The line on which the car is located

All the sprites are displayed on screen, and the game is 100% functional, with no bugs.

Co-Authored-By: EnzoGuillouche <145991192+EnzoGuillouche@users.noreply.github.com>
Co-Authored-By: Evan-UHRING <146000775+Evan-UHRING@users.noreply.github.com>